requirements/base.txt
Django>=5.1,<6
djangorestframework>=3.15,<4
djangorestframework-simplejwt>=5.3,<6
django-filter>=24,<26
django-environ>=0.11,<1
psycopg[binary]>=3.2,<4
